refactor: moved mocks to own file
